employee_backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
djongo==1.7.3
python-dotenv==1.0.0